+++ CHANGELOG.md
@@ -4,14 +4,18 @@ All notable changes to the electron-react-app (ERA) are listed here.
 
 <br>
 
-### v0.2.0 
+### v0.2.1
+
+- disable localhost sign in option if a local server is not running
+
+### v0.2.0
 
 - position pill above dock
 - barebones hey ito functionality
 - update versioning system
 - additional analytics
-- add readme and license 
+- add readme and license
 
 ### v0.1.0
 
-- Initial release 
+- Initial release